Added 2 maps, the outpost UI aint working, check description
There seems to be a problem on the outpost UI with it not showing the indicators for how many there are left to capture.

Also, in regards to the catapult gamemode, if a player loses all their units, the game breaks. need to make it so that if you lose all your units, the game ends. AI will be excluded from this